+++ src/View/Components/Breadcrumb.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace InternetGuru\LaravelCommon\View\Components;
+
+use Illuminate\View\Component;
+use Illuminate\View\View;
+use InternetGuru\LaravelCommon\Support\Helpers;
+
+class Breadcrumb extends Component
+{
+    public string $divider;
+
+    public array $items = [];
+
+    /**
+     * Create the component instance.
+     */
+    public function __construct(string $divider = '›')
+    {
+        $this->divider = $divider;
+        $this->items = Helpers::parseUrlPath();
+    }
+
+    /**
+     * Get the view / contents that represent the component.
+     */
+    public function render(): View
+    {
+        return view('package::components.breadcrumb', [
+            'divider' => $this->divider,
+            'items' => $this->items,
+        ]);
+    }
+}